requirements.txt
pandas >= 2.0.0
python-docx 
pypdf 
pyarrow 
openpyxl 
beautifulsoup4
pillow 
opencv-python
easyocr
torch 
torchvision 
torchaudio 
#--index-url https://download.pytorch.org/whl/cu118